% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_geoboundaries.R
\name{get_admin_geoboundaries}
\alias{get_admin_geoboundaries}
\title{Get admin boundaries from GeoBoundaries via rgeoboundaries for drawing maps .}
\usage{
get_admin_geoboundaries(
  country_name = "kenya",
  boundary_type = "sscgs",
  admin_level
)
}
\arguments{
\item{country_name}{The country name of interest in lower case letters e.g. "kenya" as character string
Default value is "kenya" for Kenya.}

\item{boundary_type}{GeoBoundaries allow a choice between Simplified Single Country Unstandardized (sscu)  or
Simplified Single Country Globally Standardized (sscgs) . Character string. Default is "sscgs"}

\item{admin_level}{Character string selecting the admin level for which to obtain boundaries.
Depends on data available from GeoBoundaries. Character string. e.g. "adm1", "adm2", etc.}
}
\value{
An sf object containing shape data for \code{country_name} \code{admin_level}
}
\description{
Get admin boundaries from GeoBoundaries via rgeoboundaries for drawing maps .
}
\examples{
ken_adm3 <- get_admin_geoboundaries(country_name = "kenya",
boundary_type = "sscgs", admin_level = "adm3")
}
\author{
Anelda van der Walt
}
