% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pregnancy_data.R
\name{get_pregnancy_data}
\alias{get_pregnancy_data}
\title{Get pregnancy data provided by WiGISKe -
http://wigis.co.ke/project/visualizing-teenage-pregnancy-and-related-factors/}
\usage{
get_pregnancy_data(csv_file)
}
\arguments{
\item{csv_file}{A link to a CSV file online or a path to a local CSV file as character string.}
}
\value{
A tibble containing a cleaned up version of the pregnancy data
}
\description{
Get pregnancy data provided by WiGISKe -
http://wigis.co.ke/project/visualizing-teenage-pregnancy-and-related-factors/
}
\examples{
ken_preg <- get_pregnancy_data(csv_file = "https://tinyurl.com/y35htfoj")
}
\author{
Anelda van der Walt
}
