% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_adm_boundaries.R
\name{join_adm_boundaries}
\alias{join_adm_boundaries}
\title{Add admin boundaries to dataframe and convert to an sf object to use in mapping and other analysis}
\usage{
join_adm_boundaries(adm_column = "orgunitlevel1", boundaries_sf, data_df)
}
\arguments{
\item{adm_column}{Name of the column in dataframe/tibble that contains the admin level names.
Default: orgunitlevel1}

\item{boundaries_sf}{SF object containing the admin boundaries for the relevant admin level.
Obtained by running get_admin_geoboundaries()}

\item{data_df}{Dataframe or tibble containing the data that needs to be converted to sf object. Obtained
by running get_pregnancy_data() and normalise_pregnancy_data()}
}
\value{
An sf object
}
\description{
Add admin boundaries to dataframe and convert to an sf object to use in mapping and other analysis
}
\author{
Anelda van der Walt
}
