% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_norm_preg.R
\name{map_norm_preg}
\alias{map_norm_preg}
\title{Draw cloropleth map of normalised adolescent pregnancy data
Can select age group}
\usage{
map_norm_preg(data_sf, age_group, year)
}
\arguments{
\item{data_sf}{The dataframe/tibble that contains normalised pregnancy data and
geometry data for the admin boundary of interest- created by the functions normalise_pregnancy_data()
and join_adm_boundaries(). Must be an sf object.}

\item{age_group}{Either 1014 for ages 10 - 14 or 1519 for ages 15 - 19}

\item{year}{The year for which the map should be drawn (can be 2016, 2017, 2018, 2019 or 2020)}
}
\value{
A cloropleth map showing occurrance of pregnancies for a specific age group across the counties in Kenya
}
\description{
Draw cloropleth map of normalised adolescent pregnancy data
Can select age group
}
\seealso{
normalise_pregnancy_data join_adm_boundaries
}
\author{
Anelda van der Walt
}
