% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_pregnancy_data.R
\name{normalise_pregnancy_data}
\alias{normalise_pregnancy_data}
\title{Use World bank data to normalise number of female teenagers from 2016 - 2020}
\usage{
normalise_pregnancy_data(preg_df, pop_f_1014, pop_f_1519)
}
\arguments{
\item{preg_df}{The dataframe or tibble created when running get_pregnancy_data("https://tinyurl.com/y35htfoj").}

\item{pop_f_1014}{The World Bank data for females from Kenya aged 10 - 14
obtained when running get_wb_gender_age_pop_data(indicator_code = "SP.POP.1014.FE")}

\item{pop_f_1519}{The World Bank data for females from Kenya aged 15 - 19
obtained when running get_wb_gender_age_pop_data(indicator_code = "SP.POP.1519.FE")}
}
\value{
A dataframe containing normalised pregnancy data.
}
\description{
Use World bank data to normalise number of female teenagers from 2016 - 2020
}
\examples{
norm_preg <- normalise_pregnancy_data(preg_df, pop_f_1014, pop_f_1519)
}
\seealso{
get_pregnancy_data get_wb_gender_age_pop_data
}
\author{
Anelda van der Walt
}
