% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb_gender_age_pop_data.R
\name{get_wb_gender_age_pop_data}
\alias{get_wb_gender_age_pop_data}
\title{Get World Bank population data by gender and age group for several years.}
\usage{
get_wb_gender_age_pop_data(country_iso, indicator_code, start, end, new_date)
}
\arguments{
\item{country_iso}{The three letter ISO code for a country in capital letters as character string.}

\item{indicator_code}{The World Bank data indicator code in all capitals for the dataset in question
- can be obtained by running wb_indicators() as character string e.g. SP.POP.1014.FE .}

\item{start}{The start year from which to collect data from the original dataset - numerical}

\item{end}{The end year until which to collect data from the original dataset - numerical}

\item{new_date}{The new year for which to calculate population data based on existing years
in the dataset - as numerical}
}
\value{
A dataframe containing population data from \code{start} to \code{end} for the
indicator \code{indicator_code} with an additional row for the new year.
}
\description{
Get World Bank population data by gender and age group for several years.
}
\examples{
get_wb_gender_age_pop_data(country_iso = "KEN", indicator_code = "SP.POP.1014.FE",
start = 2016, end = 2019, new_date = 2020)
}
\author{
Anelda van der Walt
}
